Connect the js files
