+++ Color.js
@@ -1,22 +1,18 @@
-// var Palette = require("Palette.js")
-
-// class Color {
-//     constructor() {
-//         this.locked = false;
-//         this.hexCode = this.createHexCode();
-//     }
-//     createHexCode() {
-//       let hexValues = ["A", "B", "C", "D", "E", "F", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9"];
-//       let hexArray = [];
-//         for (var i = 0; i < 6; i++) {
-//             var randomIndex = getRandomIndex(hexValues);
-//             var randomValue = hexValues[randomIndex];
-//             hexArray.push(randomValue);
-//             var commaSeparated = hexArray.join("")
-//             commaSeparated = `#${commaSeparated}`
-//         }
-//         return commaSeparated
-//     }
-// }
-
-// module.exports = Color 
+class Color {
+    constructor() {
+        this.locked = false;
+        this.hexCode = this.createHexCode();
+    }
+    createHexCode() {
+      let hexValues = ["A", "B", "C", "D", "E", "F", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9"];
+      let hexArray = [];
+        for (var i = 0; i < 6; i++) {
+            var randomIndex = getRandomIndex(hexValues);
+            var randomValue = hexValues[randomIndex];
+            hexArray.push(randomValue);
+            var commaSeparated = hexArray.join("")
+            commaSeparated = `#${commaSeparated}`
+        }
+        return commaSeparated
+    }
+}
