+++ Palette.js
@@ -1,19 +1,15 @@
-// var Color = require("Color.js")
+class Palette {
+    constructor() {
+      this.colors = [new Color(), new Color(), new Color(), new Color(), new Color()]
+      this.id = Date.now()
+    }
 
-// class Palette {
-//     constructor() {
-//       this.colors = [new Color(), new Color(), new Color(), new Color(), new Color()]
-//       this.id = Date.now()
-//     }
-  
-//     changeColors() {
-//       console.log("this.id", this.id) 
-//       for (var i=0; i < this.colors.length; i++) {
-//         if (this.colors[i].locked === false) {  
-//           this.colors[i] = new Color();
-//         }  
-//       }
-//     }
-//   }
-
-//   module.exports = Palette 
+    changeColors() {
+      console.log("this.id", this.id)
+      for (var i=0; i < this.colors.length; i++) {
+        if (this.colors[i].locked === false) {
+          this.colors[i] = new Color();
+        }
+      }
+    }
+  }
