+++ index.html
@@ -64,8 +64,8 @@
         </div>
       </div>
     </div>
-    <script src="scripts.js"></script>
-    <!-- <script src="Color.js"></script>
-    <script src="Palette.js"></script> -->
+    <script type="text/javascript" src="Color.js"></script>
+    <script type="text/javascript" src="Palette.js"></script>
+    <script type="text/javascript" src="scripts.js"></script>
   </body>
 </html>
